telemetry-service/pyproject.toml
[project]
name = "telemetry-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "debugpy>=1.8.16",
    "fastapi[standard]>=0.116.1",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.24",
]
